refactor(panel): merge panelitem into panel and add bug report
- Merges the functionality of the `!panelitem` command into the `!panel` command.
- The `!panel` command now gives the player the enchanted panel item.
- The permission level for `!panel` has been set to 1 (admin and above).
- The redundant `!panelitem` command has been deleted.
- Adds a new bug fix report in `Dev/Bug_Fix_Report.md` detailing the investigation and resolution of previous issues with the `panelitem` command.
