+++ AntiCheatsBP/scripts/modules/commands/help.js
@@ -30,7 +30,7 @@ function showCategorizedHelp(player, userPermissionLevel) {
         },
         '§cModeration': {
             systemToggle: null,
-            commands: ['panel', 'panelitem', 'kick', 'mute', 'unmute', 'freeze', 'vanish', 'invsee', 'tp', 'ban', 'unban', 'clearchat']
+            commands: ['panel', 'kick', 'mute', 'unmute', 'freeze', 'vanish', 'invsee', 'tp', 'ban', 'unban', 'clearchat']
         },
         '§9Administration': {
             systemToggle: null,
