+++ Dev/Bug_Fix_Report.md
@@ -1,0 +1,31 @@
+# Bug Fix Report
+
+This report details the investigation and resolution of several issues encountered during the development of the bounty system and the `!panelitem` command.
+
+## 1. `!panelitem` Command Crash & Loading Errors
+
+**Issue:** The `!panelitem` command was causing issues, first crashing the server when executed, and then failing to load altogether.
+
+**Investigation:**
+-   **Initial Crash:** The error log pointed to the line where the enchantment was being added to the item. I discovered that the item's definition in `AntiCheatsBP/items/panel.json` was missing the `minecraft:enchantable` component.
+-   **Item Parsing Error:** My first attempt to fix this by adding the component with `slot: "mainhand"` resulted in an item parsing error. I found through documentation that `"mainhand"` is not a valid slot type and corrected it to `"sword"`.
+-   **Loading Error:** After fixing the parsing error, the command started to fail to load. I suspected an issue with how the enchantment type was being referenced. I tried using `new EnchantmentType('curse_of_vanishing')`, `MinecraftEnchantmentTypes.vanishing`, and finally `new EnchantmentType('vanishing')`. The use of `MinecraftEnchantmentTypes` seemed to be incompatible with the current scripting environment, causing the loading error.
+
+**Resolution:**
+1.  Added the `minecraft:enchantable` component to `AntiCheatsBP/items/panel.json` with a valid slot type (`"sword"`).
+2.  Used `new EnchantmentType('vanishing')` to get the enchantment type in `panelitem.js`, which proved to be the most stable solution.
+
+## 2. General Command Loading Errors
+
+**Issue:** At one point, several new command files (`bounty.js`, `listbounty.js`, `rbounty.js`) were failing to load.
+
+**Investigation:**
+-   The error logs were not descriptive.
+-   I initially suspected incorrect import paths for the `commandManager`. After correcting these, the issue persisted.
+-   Further investigation revealed that the files themselves were inexplicably missing from the repository, likely due to an environment issue.
+
+**Resolution:**
+1.  The primary fix was to recreate the missing command files from my development history.
+2.  I also ensured that all commands use the correct relative import path for the local `commandManager` (`./commandManager.js`).
+
+These steps have resolved the reported issues and stabilized the new features.
